fix: added delete reimbursement code during delete bank account

+++ src/libs/actions/BankAccounts.js
@@ -103,6 +103,12 @@ function addPersonalBankAccount(account, password, plaidLinkToken) {
  * @param {Number} bankAccountID
  */
 function deleteBankAccount(bankAccountID) {
+    const reimbursementBankAccountId = lodashGet(store.getReimbursementAccountInSetup(), 'bankAccountID');
+    if(reimbursementBankAccountId === bankAccountID) {
+        BankAccounts.resetFreePlanBankAccount();
+        return;
+    }
+
     API.DeleteBankAccount({
         bankAccountID,
     }).then((response) => {
